+++ CHANGELOG.md
@@ -6,3 +6,10 @@
 
 - 发布第一个公开版本，包含 icon、button、checkbox、radio、switch、slider、input-number、datepicker、select、tabs、breadcrumb、collapse、pagination、modal、message、popover、progress 等组件。
 - 发布 [Bee Mobile 首页](https://bee-mobiles.github.io) 和入门文档。
+
+## 0.1.1(2018-04-03)
+
+**CHANGES**
+
+- 修复找不到Indicator、MessageBox、Notification、Toast模块。
+- <img className="emoji" alt="tada" height="20" width="20" src="https://assets-cdn.github.com/images/icons/emoji/unicode/1f389.png"/>发布create-bee-mobile脚手架，可快速搭建基于Webpack构建的React项目。